refactor: 주문 요약 상세 ui 컴포넌트 분리
